fix: move reliability modal to Modals directory + update infrastructure docs
